Module rb_soft::api

Higher level functions, to deal with events/binary reprentation of it, configure the drs4, etc.

Structs

Commander

Deal with incoming commands

FIXME

Constants

DMA_RESET_TRIES 🔒

HEARTBEAT

SAVE_RESTART_TRIES 🔒

SLEEP_AFTER_REG_WRITE 🔒

Functions

buff_handler

Deal with the raw data buffers.

debug_and_ok 🔒

! Little helper

event_cache

Recieve the events and hold them in a cache until they are requested

event_payload_worker

Transforms raw bytestream to RBEventPayload

get_buff_size

Get the blob buffer size from occupancy register

heartbeat 🔒

! Send out a signal periodically to all the threads. If they don’t answer in timely manner, call a doctor.

kill_run 🔒

Make sure a run stops

make_sure_it_runs 🔒

Ensure the buffers are filled and everything is prepared for data taking

reset_data_memory_aggressively 🔒

! Somehow, it is not always successful to reset the DMA and the data buffers. Let’s try an aggressive scheme and do it several times. If we fail, something is wrong and we panic

runner

A simple routine which runs until a certain amoutn of events are acquired

setup_drs4

Prepare the whole readoutboard for data taking.