Module rb_soft::control

Functions

blob_buffer_reset

! Reset means, the memory can be used again, but it does not mean it clears the memory.

clear_dma_memory

! Reset the DMA memory (blob data) and write 0s

get_blob_buffer_occ

! Get the blob buffer occupancy for one of the two buffers

get_board_id

! Get the board ID from the control registers.

get_device_dna

! The device DNA is a unique identifier

get_dma_pointer

! FIXME

get_event_count

! Get the event counter from the DRS4

get_lost_event_count

! Get the lost events event counter from the DRS4

get_lost_trigger_rate

! Get the rate of the lost triggers by the DRS4

get_trigger_rate

! Get the recorded triggers by the DRS4

idle_drs4_daq

! Put the daq in idle state, that is stop data taking

reset_daq

reset_dma

! Resets the DMA state machine.

reset_drs_event_ctr

! Reset of the internal event counter

set_drs4_configure

! This simply sets the configure bit.

set_master_trigger_mode

set_readout_all_channels_and_ch9

! Enable the readout of all channels + the 9th channel

start_drs4_daq

Start DRS4 data acquistion

switch_ram_buffer

! Toggle between the data buffers A and B