Module rb_soft::memory

Structs

RegisterError

Enums

BlobBuffer

! There are 2 data buffers, commonly denoted as “A” and “B”. A -> /dev/uio1 B -> /dev/uio2

Constants

SIZEOF_U32

The size of a 32bit unsigned int in byte (all words in registers are u32)

UIO0

memory locations for the control registers /dev/uio0 - DRS4 control /dev/uio1 - buffer 1 for blobs /dev/uio2 - buffer 2 for blobs

UIO1

UIO1_MAX_OCCUPANCY

UIO1_MIN_OCCUPANCY

Data buffer related constants The data buffer is /dev/uio1 and /dev/uio2 are internally a single buffer but with 2 halves.

UIO2

UIO2_MAX_OCCUPANCY

UIO2_MIN_OCCUPANCY

Functions

dump_mem 🔒 Deprecated

For debugging. This just prints the memory at a certain address

map_physical_mem_read

Allow READ access to the memory registers at /dev/uio**

map_physical_mem_write

Allow WRITE access to the memory registers at /dev/uio0

read_control_reg

! Get a single value from a 32bit (1 word) register This reads ONLY control registers (in /dev/uio0)

read_data_buffer

! Read one of the data buffers and return a bytestream from the given address with the length in events.

write_control_reg